pyproject.toml
[project]
name = "locust-performace-poc"
version = "0.1.0"
description = "Performance testing framework built using Locust with structured API abstraction and data-driven login."
authors = [
    {name = "priyanshi.saxena@thoughtworks.com"}
]
requires-python = ">=3.14"
dependencies = [
    "locust (>=2.43.3,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry]
package-mode = false